revise build script

+++ build-production-containers.sh
@@ -6,23 +6,46 @@ if [ "$#" != 1 ] ; then
 fi
 VERSION="${1}"
 
-echo "Building API container"
-docker build --rm -t utseresearch/describo-online-api:latest -f Dockerfile.api-build .
-docker tag utseresearch/describo-online-api:latest utseresearch/describo-online-api:${VERSION}
-docker rmi $(docker images | grep none | awk '{print $3}')
+read -p '>> Build the API code? [y|N] ' resp
+if [ "$resp"  == "y" ] ; then
+    cd api
+    npm run build:production
+    cd -
+    echo
+fi
 
-echo "Building UI code"
-docker run --rm -v $PWD/ui:/srv/ui -v $PWD/../describo-ui-plugins:/srv/plugins -w /srv/ui node:14-buster npm run build
+read -p '>> Build the UI code? [y|N] ' resp
+if [ "$resp"  == "y" ] ; then
+    cd ui
+    npm run build
+    cd -
+    echo
+fi
 
-echo "Building UI container"
-docker build --rm -t utseresearch/describo-online-ui:latest -f Dockerfile.ui-build .
-docker tag utseresearch/describo-online-ui:latest utseresearch/describo-online-ui:${VERSION}
+read -p '>> Build the containers? [y|N] ' resp
+if [ "$resp" == "y" ] ; then
+    echo "Building API container"
+    docker build --rm -t arkisto/describo-online-api:latest -f Dockerfile.api-build .
+    docker tag arkisto/describo-online-api:latest arkisto/describo-online-api:${VERSION}
+    docker rmi $(docker images | grep none | awk '{print $3}')
 
-docker rmi $(docker images | grep none | awk '{print $3}')
+    echo "Building UI code"
+    docker run --rm -v $PWD/ui:/srv/ui -v $PWD/../describo-ui-plugins:/srv/plugins -w /srv/ui node:14-buster npm run build
 
-echo "Pushing built containers to docker hub"
-docker login
-docker push utseresearch/describo-online-api:latest
-docker push utseresearch/describo-online-api:${VERSION}
-docker push utseresearch/describo-online-ui:latest
-docker push utseresearch/describo-online-ui:${VERSION}
+    echo "Building UI container"
+    docker build --rm -t arkisto/describo-online-ui:latest -f Dockerfile.ui-build .
+    docker tag arkisto/describo-online-ui:latest arkisto/describo-online-ui:${VERSION}
+
+    docker rmi $(docker images | grep none | awk '{print $3}')
+    echo
+fi
+
+read -p '>> Push the containers to docker hub? [y|N] ' resp
+if [ "$resp" == "y" ] ; then
+    echo "Pushing built containers to docker hub"
+    docker login
+    docker push arkisto/describo-online-api:latest
+    docker push arkisto/describo-online-api:${VERSION}
+    docker push arkistor/describo-online-ui:latest
+    docker push arkisto/describo-online-ui:${VERSION}
+fi
